pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "n2znd"
version = "0.1.0"
description = "Neural Network Zhu–Nakamura Dynamics toolkit."
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.23",
    "PyYAML>=6.0",
    "ase>=3.22",
    "nequip>=0.6",
]

[project.optional-dependencies]
tools = [
    "matplotlib>=3.8",
    "Pillow>=10.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
